+++ NEWS.md
@@ -8,7 +8,7 @@ Version 0.5.3
 
 ### Bug fixes
 
-* Fixed an existing bug with dark mode on macOS and Windows causing scrollbars to be forcibly rendered in a light color whenever a non-default preview background color is set.
+* Fixed an existing bug with dark mode or dark system color schemes on all platforms causing scrollbars to be forcibly rendered in a light color whenever a non-default preview background color is set.
 
 ### Other changes